Consolidate order input header
