--- src/components/Trade/Range/RangeHeader/RangeHeader.module.css
@@ -1,26 +1,0 @@
-.settings_container svg {
-    margin: 0 5px;
-}
-
-.token_info {
-    cursor: pointer;
-    font-size: var(--header1-size);
-    line-height: var(--header1-lh);
-    letter-spacing: -0.02em;
-    font-weight: 300;
-    color: var(--text1);
-}
-.hoverable_icon{
-    fill: var(--text2) !important;
-    color: var(--text2) !important
-}
-.hoverable_icon:hover{
-    fill: var(--accent1) !important
-}
-.settings_icon,
-.share_button {
-    cursor: pointer;
-}
-.share_button:hover {
-    color: var(--accent1) !important;
-}
